rename dir to single_agent_tests
